requirements.txt
appconfig==0.1
appdirs==1.4.3
cachetools==2.0.1
certifi==2018.4.16
chardet==3.0.4
defusedxml==0.5.0
Django==2.0.4
django-braces==1.12.0
django-cors-headers==2.2.0
django-oauth-toolkit==1.0.0
django-oauth2-provider==0.2.6.1
django-rest-framework-social-oauth2==1.1.0
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
google-api-python-client==1.6.6
google-auth==1.4.1
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.2.0
gunicorn==19.7.1
httplib2==0.11.3
idna==2.6
mysqlclient==1.3.12
oauth2client==4.1.2
oauthlib==2.0.7
Pillow==5.1.0
pyasn1==0.4.2
pyasn1-modules==0.2.1
PyJWT==1.6.1
python3-openid==3.1.0
pytz==2018.3
requests==2.18.4
requests-oauthlib==0.8.0
rsa==3.4.2
shortuuid==0.5.0
six==1.11.0
social-auth-app-django==2.1.0
social-auth-core==1.7.0
uritemplate==3.0.0
urllib3==1.22
